refactor(Operation): move transform property mutators
The Transform Property Mutators have now been moved into the
`Data/Operation` namespace as this is a more appropriate location.
