+++ Scripts/Data/Operation/TransformEulerRotationMutator.cs.meta
