--- Scripts/Mutation.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: 59e7e8dc3974eb1428330fb932ea0954
-folderAsset: yes
-DefaultImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
