--- Scripts/Mutation/TransformProperty.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: aaf2e218994952f4789322427945a876
-folderAsset: yes
-DefaultImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
