--- Tests/Editor/Mutation.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: 126e41ba0fb783147b063788fcdc2e31
-folderAsset: yes
-DefaultImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
