--- Tests/Editor/Mutation/TransformProperty.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: 252a5f30738c9c84c88371c7bc5fa57b
-folderAsset: yes
-DefaultImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
